building

+++ LycorisNet/Genome.go
@@ -11,13 +11,13 @@ type gen struct {
 }
 
 type ome struct {
-	weight        float64
+	weight        float32
 	isEnable      bool
 	innovationNum int
 }
 
 // Create new connections with full arguments.
-func newGenome(in int, out int, weight float64, isEnable bool, innovationNum int) (*gen, *ome) {
+func newGenome(in int, out int, weight float32, isEnable bool, innovationNum int) (*gen, *ome) {
 	var g = &gen{in, out}
 	var o = &ome{weight, isEnable, innovationNum}
 	return g, o
